requirements.txt
google-cloud-aiplatform
google-cloud-bigquery
python-dotenv
